Split out IsIN and ColumnStatsData tests in to their own files, as the column stats test files are getting massive
