+++ cpp/arcticdb/CMakeLists.txt
@@ -1041,7 +1041,9 @@ if(${TEST})
             pipeline/test/test_frame_allocation.cpp
             pipeline/test/test_rollback.cpp
             pipeline/test/test_value.cpp
+            pipeline/test/test_column_stats_data.cpp
             pipeline/test/test_column_stats_dispatch.cpp
+            pipeline/test/test_column_stats_isin.cpp
             util/test/test_regex.cpp
             processing/test/test_arithmetic_type_promotion.cpp
             processing/test/test_clause.cpp
