+++ cpp/arcticdb/pipeline/test/test_column_stats_dispatch.cpp
@@ -1,8 +1,6 @@
 #include <gtest/gtest.h>
 #include <arcticdb/pipeline/column_stats_dispatch.hpp>
-#include <arcticdb/pipeline/value_set.hpp>
 #include <arcticdb/pipeline/column_stats_filter.hpp>
-#include <arcticdb/pipeline/column_stats.hpp>
 #include <arcticdb/util/constants.hpp>
 
 #include <cmath>
@@ -1463,393 +1461,4 @@ INSTANTIATE_TEST_SUITE_P(
                 std::make_tuple(OperationType::EQ, StatsComparison::NONE_MATCH),
                 std::make_tuple(OperationType::NE, StatsComparison::ALL_MATCH)
         )
-);
-
-// Helper to build a column stats SegmentInMemory with the expected schema:
-//   col 0: "start_index" (NANOSECONDS_UTC64)
-//   col 1: "end_index"   (NANOSECONDS_UTC64)
-//   col 2: "v1.0_MIN(price)" (INT64)
-//   col 3: "v1.0_MAX(price)" (INT64)
-//
-// Each entry in |rows| is (start_index, end_index, min_price, max_price).
-// If a row's start_index is std::nullopt, that row has an absent start_index
-struct StatsSegmentRow {
-    std::optional<timestamp> start_index;
-    std::optional<timestamp> end_index;
-    int64_t min_price;
-    int64_t max_price;
-};
-
-SegmentInMemory build_stats_segment(const std::vector<StatsSegmentRow>& rows) {
-    auto start_col = std::make_shared<Column>(make_scalar_type(DataType::NANOSECONDS_UTC64), Sparsity::PERMITTED);
-    auto end_col = std::make_shared<Column>(make_scalar_type(DataType::NANOSECONDS_UTC64), Sparsity::PERMITTED);
-    auto min_col = std::make_shared<Column>(make_scalar_type(DataType::INT64), Sparsity::PERMITTED);
-    auto max_col = std::make_shared<Column>(make_scalar_type(DataType::INT64), Sparsity::PERMITTED);
-
-    for (const auto& row : rows) {
-        if (row.start_index.has_value()) {
-            start_col->push_back<timestamp>(*row.start_index);
-        } else {
-            start_col->mark_absent_rows(1);
-        }
-        if (row.end_index.has_value()) {
-            end_col->push_back<timestamp>(*row.end_index);
-        } else {
-            end_col->mark_absent_rows(1);
-        }
-        min_col->push_back<int64_t>(row.min_price);
-        max_col->push_back<int64_t>(row.max_price);
-    }
-
-    auto last_row = static_cast<ssize_t>(rows.size()) - 1;
-    start_col->set_row_data(last_row);
-    end_col->set_row_data(last_row);
-    min_col->set_row_data(last_row);
-    max_col->set_row_data(last_row);
-
-    SegmentInMemory seg;
-    seg.descriptor().set_index(IndexDescriptorImpl(IndexDescriptorImpl::Type::ROWCOUNT, 0));
-    seg.add_column(scalar_field(DataType::NANOSECONDS_UTC64, start_index_column_name), start_col);
-    seg.add_column(scalar_field(DataType::NANOSECONDS_UTC64, end_index_column_name), end_col);
-    seg.add_column(scalar_field(DataType::INT64, "v1.0_MIN(price)"), min_col);
-    seg.add_column(scalar_field(DataType::INT64, "v1.0_MAX(price)"), max_col);
-    seg.set_row_data(last_row);
-    return seg;
-}
-
-// When all rows have valid index values, find_stats should return the correct stats for each row.
-TEST(ColumnStatsData, FindStatsAllRowsPresent) {
-    auto seg = build_stats_segment({
-            {100, 200, 10, 20},
-            {300, 400, 30, 40},
-            {500, 600, 50, 60},
-    });
-
-    ColumnStatsData data(std::move(seg));
-    ASSERT_FALSE(data.empty());
-
-    auto* row0 = data.find_stats(100, 200);
-    ASSERT_NE(row0, nullptr);
-    auto it0 = row0->stats_for_column.find("price");
-    ASSERT_NE(it0, row0->stats_for_column.end());
-    ASSERT_EQ(it0->second.min->get<int64_t>(), 10);
-    ASSERT_EQ(it0->second.max->get<int64_t>(), 20);
-
-    auto* row2 = data.find_stats(500, 600);
-    ASSERT_NE(row2, nullptr);
-    auto it2 = row2->stats_for_column.find("price");
-    ASSERT_NE(it2, row2->stats_for_column.end());
-    ASSERT_EQ(it2->second.min->get<int64_t>(), 50);
-    ASSERT_EQ(it2->second.max->get<int64_t>(), 60);
-}
-
-// A malformed row (absent start_index) causes the entire ColumnStatsData to be discarded,
-// even if valid rows were already parsed.
-TEST(ColumnStatsData, MalformedMiddleRowDiscardsAll) {
-    auto seg = build_stats_segment({
-            {100, 200, 10, 20},          // valid
-            {std::nullopt, 400, 30, 40}, // malformed — absent start_index
-            {500, 600, 50, 60},          // never reached
-    });
-
-    ColumnStatsData data(std::move(seg));
-    ASSERT_TRUE(data.empty());
-    ASSERT_EQ(data.find_stats(100, 200), nullptr);
-    ASSERT_EQ(data.find_stats(500, 600), nullptr);
-}
-
-// Absent end_index also triggers discard.
-TEST(ColumnStatsData, MalformedEndIndexDiscardsAll) {
-    auto seg = build_stats_segment({
-            {100, std::nullopt, 10, 20},
-    });
-
-    ColumnStatsData data(std::move(seg));
-    ASSERT_TRUE(data.empty());
-}
-
-// Lookup with non-existent indices returns nullptr.
-TEST(ColumnStatsData, FindStatsNonExistentIndex) {
-    auto seg = build_stats_segment({
-            {100, 200, 10, 20},
-    });
-
-    ColumnStatsData data(std::move(seg));
-    ASSERT_EQ(data.find_stats(999, 888), nullptr);
-}
-
-// An empty segment produces empty ColumnStatsData.
-TEST(ColumnStatsData, EmptySegment) {
-    auto seg = build_stats_segment({});
-
-    ColumnStatsData data(std::move(seg));
-    ASSERT_TRUE(data.empty());
-    ASSERT_EQ(data.find_stats(0, 0), nullptr);
-}
-
-// Two row-slices with the same (start_index, end_index) but different min/max stats.
-// This can happen with timestamp-indexed symbols when two segments span identical timestamp ranges
-// (e.g. segments where all rows share the same timestamp).
-// Both entries are dropped from the lookup so that find_stats returns nullptr, forcing the
-// segments to be read without pruning rather than using the wrong stats.
-TEST(ColumnStatsData, DuplicateIndexPairDropsBothRows) {
-    auto seg = build_stats_segment({
-            {100, 200, 10, 20}, // row 0: price in [10, 20]
-            {100, 200, 50, 60}, // row 1: same index range, price in [50, 60]
-    });
-
-    ColumnStatsData data(std::move(seg));
-    ASSERT_FALSE(data.empty());
-
-    // Neither row is reachable via find_stats because the key is ambiguous.
-    ASSERT_EQ(data.find_stats(100, 200), nullptr);
-}
-
-// Duplicate keys are dropped but non-duplicate keys in the same segment are unaffected.
-TEST(ColumnStatsData, DuplicateIndexPairDoesNotAffectOtherRows) {
-    auto seg = build_stats_segment({
-            {100, 300, 10, 20}, // row 0: unique key
-            {300, 400, 30, 40}, // row 1: duplicate key (first)
-            {300, 400, 50, 60}, // row 2: duplicate key (second)
-            {400, 600, 70, 80}, // row 3: unique key
-    });
-
-    ColumnStatsData data(std::move(seg));
-    ASSERT_FALSE(data.empty());
-
-    // Unique rows are still reachable.
-    auto* row0 = data.find_stats(100, 300);
-    ASSERT_NE(row0, nullptr);
-    ASSERT_EQ(row0->stats_for_column.at("price").min->get<int64_t>(), 10);
-    ASSERT_EQ(row0->stats_for_column.at("price").max->get<int64_t>(), 20);
-
-    auto* row3 = data.find_stats(400, 600);
-    ASSERT_NE(row3, nullptr);
-    ASSERT_EQ(row3->stats_for_column.at("price").min->get<int64_t>(), 70);
-    ASSERT_EQ(row3->stats_for_column.at("price").max->get<int64_t>(), 80);
-
-    // Duplicate key is not reachable.
-    ASSERT_EQ(data.find_stats(300, 400), nullptr);
-}
-
-namespace {
-
-template<typename T>
-std::shared_ptr<ValueSet> make_numeric_value_set(std::initializer_list<T> values) {
-    auto set = std::make_shared<std::unordered_set<T>>(values);
-    return std::make_shared<ValueSet>(NumericSetType{std::move(set)});
-}
-
-template<typename T>
-std::shared_ptr<ValueSet> make_empty_value_set() {
-    auto set = std::make_shared<std::unordered_set<T>>();
-    return std::make_shared<ValueSet>(NumericSetType{std::move(set)});
-}
-
-} // namespace
-
-TEST(StatsMembershipComparator, EmptySetIsinNoneMatch) {
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(int64_t{1})}, std::optional<Value>{construct_value(int64_t{10})}
-    };
-    auto vs = make_empty_value_set<int64_t>();
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::NONE_MATCH);
-}
-
-TEST(StatsMembershipComparator, EmptySetIsnotinAllMatch) {
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(int64_t{1})}, std::optional<Value>{construct_value(int64_t{10})}
-    };
-    auto vs = make_empty_value_set<int64_t>();
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISNOTIN), StatsComparison::ALL_MATCH);
-}
-
-TEST(StatsMembershipComparator, SetBelowBlockRange) {
-    // Block [10, 20], set {1, 5} -> set_max=5 < block_min=10 -> NONE_MATCH for ISIN
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(int64_t{10})}, std::optional<Value>{construct_value(int64_t{20})}
-    };
-    auto vs = make_numeric_value_set<int64_t>({1, 5});
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::NONE_MATCH);
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISNOTIN), StatsComparison::ALL_MATCH);
-}
-
-TEST(StatsMembershipComparator, SetAboveBlockRange) {
-    // Block [1, 5], set {10, 20} -> set_min=10 > block_max=5 -> NONE_MATCH for ISIN
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(int64_t{1})}, std::optional<Value>{construct_value(int64_t{5})}
-    };
-    auto vs = make_numeric_value_set<int64_t>({10, 20});
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::NONE_MATCH);
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISNOTIN), StatsComparison::ALL_MATCH);
-}
-
-TEST(StatsMembershipComparator, SetOverlapsBlockRange) {
-    // Block [1, 10], set {5, 15} -> overlaps -> UNKNOWN
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(int64_t{1})}, std::optional<Value>{construct_value(int64_t{10})}
-    };
-    auto vs = make_numeric_value_set<int64_t>({5, 15});
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::UNKNOWN);
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISNOTIN), StatsComparison::UNKNOWN);
-}
-
-TEST(StatsMembershipComparator, SingleElementMatchesSingleValueBlock) {
-    // Block [5, 5], set {5} -> ALL_MATCH for ISIN
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(int64_t{5})}, std::optional<Value>{construct_value(int64_t{5})}
-    };
-    auto vs = make_numeric_value_set<int64_t>({5});
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::ALL_MATCH);
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISNOTIN), StatsComparison::NONE_MATCH);
-}
-
-TEST(StatsMembershipComparator, MixedTypesInt32StatsDoubleSet) {
-    // Block [1, 10] as int32, set {15.0, 20.0} as double -> set_min=15 > block_max=10 -> NONE_MATCH
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(int32_t{1})}, std::optional<Value>{construct_value(int32_t{10})}
-    };
-    auto vs = make_numeric_value_set<double>({15.0, 20.0});
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::NONE_MATCH);
-}
-
-TEST(StatsMembershipComparator, NaNInStatsMin) {
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(std::numeric_limits<double>::quiet_NaN())},
-            std::optional<Value>{construct_value(10.0)}
-    };
-    auto vs = make_numeric_value_set<double>({5.0});
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::UNKNOWN);
-}
-
-TEST(StatsMembershipComparator, NaNInStatsMax) {
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(1.0)},
-            std::optional<Value>{construct_value(std::numeric_limits<double>::quiet_NaN())}
-    };
-    auto vs = make_numeric_value_set<double>({5.0});
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::UNKNOWN);
-}
-
-TEST(StatsMembershipComparator, MissingStats) {
-    ColumnStatsValues csv{std::nullopt, std::nullopt};
-    auto vs = make_numeric_value_set<double>({5.0});
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::UNKNOWN);
-}
-
-TEST(StatsMembershipComparator, SetContainedWithinBlockRange) {
-    // Block [1, 100], set {3, 7} -> overlap -> UNKNOWN
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(int64_t{1})}, std::optional<Value>{construct_value(int64_t{100})}
-    };
-    auto vs = make_numeric_value_set<int64_t>({3, 7});
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), StatsComparison::UNKNOWN);
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISNOTIN), StatsComparison::UNKNOWN);
-}
-
-// Per-element pruning tests: the set's min/max range overlaps the block range,
-// but individual element iteration can still prune.
-class StatsMembershipPerElementInt64Test
-    : public ::testing::TestWithParam<
-              std::tuple<int64_t, int64_t, std::vector<int64_t>, StatsComparison, StatsComparison>> {};
-
-TEST_P(StatsMembershipPerElementInt64Test, IsinAndIsnotin) {
-    auto [block_min, block_max, set_values, expected_isin, expected_isnotin] = GetParam();
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(block_min)}, std::optional<Value>{construct_value(block_max)}
-    };
-    auto vs = std::make_shared<ValueSet>(
-            NumericSetType{std::make_shared<std::unordered_set<int64_t>>(set_values.begin(), set_values.end())}
-    );
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), expected_isin);
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISNOTIN), expected_isnotin);
-}
-
-INSTANTIATE_TEST_SUITE_P(
-        RangeOverlapsAllOutside, StatsMembershipPerElementInt64Test,
-        ::testing::Values(std::make_tuple(
-                int64_t{5}, int64_t{10}, std::vector<int64_t>{1, 3, 15}, StatsComparison::NONE_MATCH,
-                StatsComparison::ALL_MATCH
-        ))
-);
-
-INSTANTIATE_TEST_SUITE_P(
-        RangeOverlapsSomeInside, StatsMembershipPerElementInt64Test,
-        ::testing::Values(std::make_tuple(
-                int64_t{5}, int64_t{10}, std::vector<int64_t>{1, 7, 15}, StatsComparison::UNKNOWN,
-                StatsComparison::UNKNOWN
-        ))
-);
-
-INSTANTIATE_TEST_SUITE_P(
-        SingleValueBlockHit, StatsMembershipPerElementInt64Test,
-        ::testing::Values(std::make_tuple(
-                int64_t{5}, int64_t{5}, std::vector<int64_t>{3, 5, 8}, StatsComparison::ALL_MATCH,
-                StatsComparison::NONE_MATCH
-        ))
-);
-
-INSTANTIATE_TEST_SUITE_P(
-        SingleValueBlockMiss, StatsMembershipPerElementInt64Test,
-        ::testing::Values(std::make_tuple(
-                int64_t{5}, int64_t{5}, std::vector<int64_t>{3, 8, 12}, StatsComparison::NONE_MATCH,
-                StatsComparison::ALL_MATCH
-        ))
-);
-
-class StatsMembershipNaNTest
-    : public ::testing::TestWithParam<
-              std::tuple<double, double, std::vector<double>, StatsComparison, StatsComparison>> {};
-
-TEST_P(StatsMembershipNaNTest, IsinAndIsnotin) {
-    auto [block_min, block_max, set_values, expected_isin, expected_isnotin] = GetParam();
-    ColumnStatsValues csv{
-            std::optional<Value>{construct_value(block_min)}, std::optional<Value>{construct_value(block_max)}
-    };
-    auto vs = std::make_shared<ValueSet>(
-            NumericSetType{std::make_shared<std::unordered_set<double>>(set_values.begin(), set_values.end())}
-    );
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISIN), expected_isin);
-    ASSERT_EQ(stats_membership_comparator(csv, *vs, OperationType::ISNOTIN), expected_isnotin);
-}
-
-INSTANTIATE_TEST_SUITE_P(
-        NaNInSetNonNaNOutside, StatsMembershipNaNTest,
-        ::testing::Values(std::make_tuple(
-                1.0, 10.0, std::vector<double>{std::numeric_limits<double>::quiet_NaN(), 20.0},
-                StatsComparison::NONE_MATCH, StatsComparison::ALL_MATCH
-        ))
-);
-
-INSTANTIATE_TEST_SUITE_P(
-        NaNInSetNonNaNInside, StatsMembershipNaNTest,
-        ::testing::Values(std::make_tuple(
-                1.0, 10.0, std::vector<double>{std::numeric_limits<double>::quiet_NaN(), 5.0}, StatsComparison::UNKNOWN,
-                StatsComparison::UNKNOWN
-        ))
-);
-
-INSTANTIATE_TEST_SUITE_P(
-        BothStatsNaNWithNaNInSet, StatsMembershipNaNTest,
-        ::testing::Values(std::make_tuple(
-                std::numeric_limits<double>::quiet_NaN(), std::numeric_limits<double>::quiet_NaN(),
-                std::vector<double>{std::numeric_limits<double>::quiet_NaN(), 5.0}, StatsComparison::NONE_MATCH,
-                StatsComparison::ALL_MATCH
-        ))
-);
-
-INSTANTIATE_TEST_SUITE_P(
-        BothStatsNaNNoNaNInSet, StatsMembershipNaNTest,
-        ::testing::Values(std::make_tuple(
-                std::numeric_limits<double>::quiet_NaN(), std::numeric_limits<double>::quiet_NaN(),
-                std::vector<double>{5.0, 20.0}, StatsComparison::NONE_MATCH, StatsComparison::ALL_MATCH
-        ))
-);
-
-INSTANTIATE_TEST_SUITE_P(
-        AllNaNValueSet, StatsMembershipNaNTest,
-        ::testing::Values(std::make_tuple(
-                1.0, 10.0, std::vector<double>{std::numeric_limits<double>::quiet_NaN()}, StatsComparison::NONE_MATCH,
-                StatsComparison::ALL_MATCH
-        ))
 );
